+++ packages/desktop/app/components/Sidebar/index.js
@@ -1,2 +1,2 @@
-export * from './ConnectedSidebar';
+export * from './config';
 export { Sidebar } from './Sidebar';