+++ packages/lan/app/routes/apiv1/suggestions.js
@@ -2,40 +2,53 @@ import { pascal } from 'case';
 import express from 'express';
 import asyncHandler from 'express-async-handler';
 import { QueryTypes } from 'sequelize';
-
 import { NotFoundError } from 'shared/errors';
-
-import { REFERENCE_TYPE_VALUES, INVOICE_LINE_TYPES } from 'shared/constants';
+import { SURVEY_TYPES, REFERENCE_TYPE_VALUES, INVOICE_LINE_TYPES } from 'shared/constants';
 
 export const suggestions = express.Router();
 
-const defaultMapper = ({ name, code, id }) => ({ name, code, id });
-
 const defaultLimit = 25;
 
-function createSuggesterRoute(endpoint, modelName, whereSql, mapper = defaultMapper) {
+const defaultMapper = ({ name, code, id }) => ({ name, code, id });
+
+function createSuggesterRoute(
+  endpoint,
+  modelName,
+  whereSql,
+  mapper = defaultMapper,
+  searchColumn = 'name',
+) {
   suggestions.get(
     `/${endpoint}`,
     asyncHandler(async (req, res) => {
       req.checkPermission('list', modelName);
       const { models, query } = req;
-      const search = (query.q || '').trim().toLowerCase();
-      if (!search) {
-        res.send([]);
-        return;
-      }
+      const searchQuery = (query.q || '').trim().toLowerCase();
 
       const model = models[modelName];
+
+      // TODO: when removing support for sqlite, just move
+      // the postgres half of this into the main query
+      const { dialect } = model.sequelize.options;
+      const positionQuery =
+        dialect === 'sqlite'
+          ? `INSTR(LOWER(${searchColumn}), LOWER(:positionMatch)) > 1`
+          : `POSITION(LOWER(:positionMatch) in LOWER(${searchColumn})) > 1`;
+
       const results = await model.sequelize.query(
         `
-      SELECT *
-      FROM "${model.tableName}"
-      WHERE ${whereSql}
-      LIMIT :limit
-    `,
+          SELECT *
+          FROM "${model.tableName}"
+          WHERE ${whereSql}
+          ORDER BY
+            ${positionQuery},
+            ${searchColumn}
+          LIMIT :limit
+        `,
         {
           replacements: {
-            search: `%${search}%`,
+            search: `%${searchQuery}%`,
+            positionMatch: searchQuery,
             limit: defaultLimit,
           },
           type: QueryTypes.SELECT,
@@ -44,8 +57,7 @@ function createSuggesterRoute(endpoint, modelName, whereSql, mapper = defaultMap
         },
       );
 
-      const listing = results.map(mapper);
-      res.send(listing);
+      res.send(results.map(mapper));
     }),
   );
 }
@@ -101,9 +113,9 @@ function createAllRecordsSuggesterRoute(endpoint, modelName, whereSql, mapper = 
 // Records will be filtered based on the whereSql parameter. The user's search term
 // will be passed to the sql query as ":search" - see the existing suggestion
 // endpoints for usage examples.
-function createSuggester(endpoint, modelName, whereSql, mapper) {
+function createSuggester(endpoint, modelName, whereSql, mapper, searchColumn) {
   createSuggesterLookupRoute(endpoint, modelName, whereSql, mapper);
-  createSuggesterRoute(endpoint, modelName, whereSql, mapper);
+  createSuggesterRoute(endpoint, modelName, whereSql, mapper, searchColumn);
 }
 
 const createNameSuggester = (endpoint, modelName = pascal(endpoint)) =>
@@ -127,7 +139,14 @@ REFERENCE_TYPE_VALUES.map(typeName =>
 createNameSuggester('department');
 createNameSuggester('location');
 createNameSuggester('facility');
-createNameSuggester('survey');
+
+createSuggester(
+  'survey',
+  'Survey',
+  `LOWER(name) LIKE LOWER(:search) AND survey_type <> '${SURVEY_TYPES.OBSOLETE}'`,
+  ({ id, name }) => ({ id, name }),
+);
+
 createSuggester(
   'invoiceLineTypes',
   'InvoiceLineType',
@@ -143,11 +162,13 @@ createSuggester(
     id,
     name: displayName,
   }),
+  'display_name',
 );
 
 createSuggester(
   'patient',
   'Patient',
-  "LOWER(first_name || ' ' || last_name) LIKE LOWER(:search) OR LOWER(cultural_name) LIKE LOWER(:search) OR LOWER(display_id) LIKE LOWER(:search)",
+  "LOWER(first_name || ' ' || last_name) LIKE LOWER(:search) OR LOWER(display_id) LIKE LOWER(:search)",
   patient => patient,
+  'first_name',
 );