+++ packages/mobile/App/models/AdministeredVaccine.ts
@@ -2,6 +2,8 @@ import { Entity, Column, ManyToOne, BeforeUpdate, BeforeInsert, RelationId } fro
 import { BaseModel } from './BaseModel';
 import { IAdministeredVaccine, InjectionSiteType } from '~/types';
 import { Encounter } from './Encounter';
+import { Location } from './Location';
+import { Department } from './Department';
 import { ScheduledVaccine } from './ScheduledVaccine';
 import { VaccineStatus } from '~/ui/helpers/patient';
 
@@ -18,9 +20,6 @@ export class AdministeredVaccine extends BaseModel implements IAdministeredVacci
 
   @Column({ type: 'varchar', nullable: true })
   injectionSite?: InjectionSiteType;
-
-  @Column({ nullable: true })
-  location?: string;
 
   @Column({ nullable: true, default: true })
   consent: boolean;
@@ -45,6 +44,24 @@ export class AdministeredVaccine extends BaseModel implements IAdministeredVacci
 
   @RelationId(({ scheduledVaccine }: AdministeredVaccine) => scheduledVaccine)
   scheduledVaccineId: string;
+
+  @ManyToOne(
+    () => Location,
+    loc => loc.administeredVaccines,
+  )
+  location: Location;
+
+  @RelationId(({ location }: AdministeredVaccine) => location)
+  locationId: string;
+
+  @ManyToOne(
+    () => Department,
+    dep => dep.administeredVaccines,
+  )
+  department: Department;
+
+  @RelationId(({ department }: AdministeredVaccine) => department)
+  departmentId: string;
 
   @BeforeInsert()
   @BeforeUpdate()