+++ packages/sync-server/app/hl7fhir/schema.js
@@ -1,11 +1,13 @@
 import * as yup from 'yup';
-import config from 'config';
 import { isArray } from 'lodash';
 
-import { decodeIdentifier, hl7PatientFields } from './utils';
+import { isValidIdentifier, getSortParameterName } from './utils';
+import { hl7PatientFields, sortableHL7PatientFields } from './hl7PatientFields';
 
-export const IDENTIFIER_NAMESPACE = config.hl7.dataDictionaries.patientDisplayId;
 const MAX_RECORDS = 20;
+
+// Explicitly set with the direction sign
+const sortableHL7BaseFields = ['-issued', 'issued'];
 
 // List of all the fixed name parameters that we support
 const baseParameters = {
@@ -17,11 +19,7 @@ const baseParameters = {
     .test(
       'is-correct-format-and-namespace',
       'subject:identifier must be in the format "<namespace>|<id>"',
-      value => {
-        if (!value) return true;
-        const [namespace, displayId] = decodeIdentifier(value);
-        return namespace === IDENTIFIER_NAMESPACE && !!displayId;
-      },
+      isValidIdentifier,
     ),
   _count: yup
     .number()
@@ -38,7 +36,7 @@ const baseParameters = {
     .required(),
   _sort: yup
     .string()
-    .oneOf(['-issued'])
+    .oneOf(sortableHL7BaseFields)
     .default('-issued')
     .required(),
   after: yup
@@ -85,7 +83,7 @@ function noUnknownValidationMessage(obj) {
   const params = Object.keys(obj.originalValue);
 
   // Return list of unknown params
-  const unknownParams = params.filter(param => param in patient.fields === false);
+  const unknownParams = params.filter(param => param in patient.query.fields === false);
   return `Unknown or unsupported parameters: ${unknownParams.join(', ')}`;
 }
 
@@ -96,6 +94,24 @@ export const patient = {
     .object({
       ...getPatientParameters(),
       ...baseParameters,
+      // Overwrite sort for patient resource
+      _sort: yup
+        .string()
+        .test('is-supported-sort', 'Unsupported or unknown parameters in _sort', value => {
+          // Sorts are separated by commas, no whitespace
+          const sorts = value.split(',');
+          // Faster to check if one is invalid
+          const isInvalid = sorts.some(sort => {
+            // Ignore base fields
+            if (sortableHL7BaseFields.includes(sort)) return false;
+            // Sort might have a "-" at the beginning, we should ignore it here
+            const parameter = getSortParameterName(sort);
+            return sortableHL7PatientFields.includes(parameter) === false;
+          });
+          return !isInvalid;
+        })
+        .default('-issued')
+        .required(),
     })
     .noUnknown(true, noUnknownValidationMessage),
 };
@@ -117,10 +133,7 @@ export const diagnosticReport = {
         .test(
           'is-correct-format-and-namespace',
           'subject:identifier must be in the format "<namespace>|<id>"',
-          value => {
-            const [namespace, displayId] = decodeIdentifier(value);
-            return namespace === IDENTIFIER_NAMESPACE && !!displayId;
-          },
+          isValidIdentifier,
         )
         .required(),
       _include: yup